- empty PASSWORD_OPTION  when KEYSTORE_PASSWORD is unsetted - empty CONFIG_OPTION on dev - empty  ALLOW_INSECURE_UNLOCK_OPTION on any BFA

+++ scripts/geth-run-node.sh
@@ -1,7 +1,5 @@
 #!/bin/bash
 
-#!/bin/bash
-
 [[ -n ${DEBUG:-} ]] && set -x
 set -Eeuo pipefail
 
@@ -21,21 +19,19 @@ echo "GETH_INSTANCE [$GETH_INSTANCE]"
 readonly REAL_GETH_INSTANCE="$(realpath "$GETH_INSTANCE")"
 echo "REAL_GETH_INSTANCE [$REAL_GETH_INSTANCE]"
 echo "DOCKER_GETH_INSTANCE [$DOCKER_GETH_INSTANCE]"
-echo "DOCKER_GETH_INSTANCE [$DOCKER_DETACHED_MODE]"
+echo "DOCKER_DETACHED_MODE [${DOCKER_DETACHED_MODE:=-d}]"
 echo "GETH_IMAGE [$GETH_IMAGE]"
 echo "GETH_DEBUG [${GETH_DEBUG:=N}]"
 
-readonly KEYSTORE_PASSWORD_FILENAME=keystore-password.txt
-rm -f "$GETH_INSTANCE/$KEYSTORE_PASSWORD_FILENAME"
-echo "$KEYSTORE_PASSWORD" > "$GETH_INSTANCE/$KEYSTORE_PASSWORD_FILENAME"
-cat "$GETH_INSTANCE/$KEYSTORE_PASSWORD_FILENAME"
-readonly PASSWORD_OPTION="--password $DOCKER_GETH_INSTANCE/$KEYSTORE_PASSWORD_FILENAME"
-
-case "$GETH_IMAGE" in
-*v1.8.27 )
-    readonly ALLOW_INSECURE_UNLOCK_OPTION="" ;;
-* ) readonly ALLOW_INSECURE_UNLOCK_OPTION="--allow-insecure-unlock" ;;
-esac
+if [[ -v KEYSTORE_PASSWORD ]]; then
+   readonly KEYSTORE_PASSWORD_FILENAME=keystore-password.txt
+   rm -f "$GETH_INSTANCE/$KEYSTORE_PASSWORD_FILENAME"
+   echo "$KEYSTORE_PASSWORD" > "$GETH_INSTANCE/$KEYSTORE_PASSWORD_FILENAME"
+   cat "$GETH_INSTANCE/$KEYSTORE_PASSWORD_FILENAME"
+   readonly PASSWORD_OPTION="--password $DOCKER_GETH_INSTANCE/$KEYSTORE_PASSWORD_FILENAME"
+else
+   readonly PASSWORD_OPTION=""
+fi
 
 case "$GETH_DEBUG" in
    Y | y | 1 | YES | yes | OK | ok | Ok )
@@ -58,17 +54,26 @@ if [[ $ENVIRONMENT == dev ]]; then
    readonly NODISCOVER_OPTION="--nodiscover"
    # ToDo: targetgaslimit deprecated => use --miner.gasprice
    readonly TARGETGASLIMIT_OPTION="--targetgaslimit 99999999999"
+   readonly CONFIG_OPTION=""
    readonly NETWORK_ID="$DEV_NETWORK_ID"
+
+   case "$GETH_IMAGE" in
+   *v1.8.27 )  readonly ALLOW_INSECURE_UNLOCK_OPTION="" ;;
+   * )         readonly ALLOW_INSECURE_UNLOCK_OPTION="--allow-insecure-unlock" ;;
+   esac
 else
+   # some BFA
    readonly MINER_OPTIONS=""
    readonly NODISCOVER_OPTION=""
    readonly TARGETGASLIMIT_OPTION=""
+   readonly CONFIG_OPTION="--config $DOCKER_GETH_INSTANCE/$NODE/config.toml"
    case "$ENVIRONMENT" in
       bfa.mainnet ) readonly NETWORK_ID="$BFA_MAINNET_NETWORK_ID" ;;
       bfa.testnet ) readonly NETWORK_ID="$BFA_TESTNET_NETWORK_ID" ;;
       * ) echo_red "ERROR: invalid ENVIRONMENT [$ENVIRONMENT]"
          exit 1
    esac
+   readonly ALLOW_INSECURE_UNLOCK_OPTION=""
 fi
 echo "NETWORK_ID [$NETWORK_ID]"
 
@@ -81,7 +86,7 @@ fi
 
 # ToDo: --gcmode archive # Asi esta en la BFA, entender lo que implica
 
-DOCKER_CMD="docker run ${DOCKER_DETACHED_MODE:-d} \
+DOCKER_CMD="docker run "$DOCKER_DETACHED_MODE" \
        --rm \
        --name "$NODE" \
        --network $DOCKER_NETWORK_ID \
@@ -98,7 +103,7 @@ DOCKER_CMD="docker run ${DOCKER_DETACHED_MODE:-d} \
        --rpcvhosts=* \
        --rpcaddr "0.0.0.0" \
        --wsaddr "0.0.0.0" \
-       --config $DOCKER_GETH_INSTANCE/$NODE/config.toml \
+       $CONFIG_OPTION \
        $TARGETGASLIMIT_OPTION \
        $NODISCOVER_OPTION \
        $MINER_OPTIONS \
@@ -109,4 +114,5 @@ DOCKER_CMD="docker run ${DOCKER_DETACHED_MODE:-d} \
 echo "running node..."
 echo $DOCKER_CMD | sed 's/\\.*//'
 $DOCKER_CMD
+
 docker port "$NODE"
